+++ mobile/android/android-components/components/service/sync-autofill/src/main/java/mozilla/components/service/sync/autofill/GeckoCreditCardsAddressesStorageDelegate.kt
@@ -4,6 +4,7 @@
 
 package mozilla.components.service.sync.autofill
 
+import kotlinx.coroutines.CompletableDeferred
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Deferred
 import kotlinx.coroutines.Dispatchers
@@ -16,10 +17,15 @@ import mozilla.components.concept.storage.CreditCardsAddressesStorageDelegate
 
 /**
  * [CreditCardsAddressesStorageDelegate] implementation.
+ *
+ * @param storage The [CreditCardsAddressesStorage] used for looking up addresses and credit cards to autofill.
+ * @param scope [CoroutineScope] for long running operations. Defaults to using the [Dispatchers.IO].
+ * @param isCreditCardAutofillEnabled callback allowing to limit [storage] operations if autofill is disabled.
  */
 class GeckoCreditCardsAddressesStorageDelegate(
     private val storage: Lazy<CreditCardsAddressesStorage>,
-    private val scope: CoroutineScope = CoroutineScope(Dispatchers.IO)
+    private val scope: CoroutineScope = CoroutineScope(Dispatchers.IO),
+    private val isCreditCardAutofillEnabled: () -> Boolean = { false }
 ) : CreditCardsAddressesStorageDelegate {
 
     override fun decrypt(encryptedCardNumber: CreditCardNumber.Encrypted): CreditCardNumber.Plaintext? {
@@ -39,6 +45,10 @@ class GeckoCreditCardsAddressesStorageDelegate(
     }
 
     override fun onCreditCardsFetch(): Deferred<List<CreditCard>> {
+        if (isCreditCardAutofillEnabled().not()) {
+            return CompletableDeferred(listOf())
+        }
+
         return scope.async {
             storage.value.getAllCreditCards()
         }